Refactor argument processing: rename process_and_push to push_if_valid, update ft_atol return type to long, and improve input validation logic

+++ src/utils/conversions_utils.c
@@ -34,7 +34,7 @@ int	determine_sign(const char *c, int *i)
 	return (sign);
 }
 
-int	ft_atol(const char *str)
+long	ft_atol(const char *str)
 {
 	long	total;
 	int		sign;
